# Conflicts:
#	src/main/webapp/html/teacher-grade-manager.html
#	src/main/webapp/js/teacher-grade-manager.js

+++ src/main/java/com/hust/bigdataplatform/controller/ExperimentController.java
@@ -28,6 +28,7 @@ import com.hust.bigdataplatform.service.ExperimentService;
 import com.hust.bigdataplatform.service.FileService;
 import com.hust.bigdataplatform.service.SessionService;
 import com.hust.bigdataplatform.util.DateConverter;
+import com.hust.bigdataplatform.util.ImageUtil;
 import com.hust.bigdataplatform.util.ResultUtil;
 import com.hust.bigdataplatform.util.UploadUtils;
 import com.hust.bigdataplatform.util.fileUtil;
@@ -156,18 +157,23 @@ public class ExperimentController {
 		experiment.setExperimentSubmitdemand(" ");
 		experiment.setExperimentCreatetime(new Date());
 		experiment.setExperimentDeadline(new DateConverter().convert(deadline));
-		experiment.setExperimentManualpath(Constant.DIRECTORY.EXPERIMENT_REPORT+uid);
-		System.out.println("实验手册："+Constant.DIRECTORY.EXPERIMENT_REPORT);
-		experiment.setExperimentVideopath(Constant.DIRECTORY.EXPERIMENT_VIDEO+uid);
-		System.out.println("实验视频："+Constant.DIRECTORY.EXPERIMENT_VIDEO);
+//		experiment.setExperimentManualpath(Constant.DIRECTORY.EXPERIMENT_REPORT+uid);
+//		System.out.println("实验手册："+Constant.DIRECTORY.EXPERIMENT_REPORT);
+//		experiment.setExperimentVideopath(Constant.DIRECTORY.EXPERIMENT_VIDEO+uid);
+//		System.out.println("实验视频："+Constant.DIRECTORY.EXPERIMENT_VIDEO);
 		experiment.setExperimentReportpath(Constant.DIRECTORY.REPORT_SUBMIT+uid);
 		experiment.setExperimentResultspath(Constant.DIRECTORY.EXPERIMENT_DATA_SUBMIT+uid);
 		int status = experimentService.addExperiment(experiment);
 		if (status==0) {
 			return ResultUtil.errorWithMsg("添加实验失败！");
 		}
+		//当添加实验后，在experimentScore表中插入记录
+		int s = experimentScoreService.AddExperimentScore(courseId, uid);
+		if (s==0) {
+			return ResultUtil.errorWithMsg("实验分数表初始化失败！");
+		}
 		//添加一个实验记录时启动自动评分线程
-		AutoRating.rating(experiment, "", experimentScoreService);
+		AutoRating.rating(experiment, "");
 		return ResultUtil.success(experiment);
 	}
 	
@@ -194,8 +200,8 @@ public class ExperimentController {
 		}
 
 		//更新实验信息时启动自动评分线程
-		AutoRating.rating(experiment.get(0), oldName, experimentScoreService);
-		return ResultUtil.errorWithMsg("修改实验内容失成功！");
+		AutoRating.rating(experiment.get(0), oldName);
+		return ResultUtil.errorWithMsg("修改实验内容成功！");
 	}
 	
 	@RequestMapping("/UpdateExpSubmitDemand")
@@ -236,13 +242,13 @@ public class ExperimentController {
 		experimentFile.setExperimentId(experimentId);
 		experimentFile.setFileId(uid);
 		
-		String road = "";
-		if (type.equals("PDF")) {
-			road=experiment.get(0).getExperimentManualpath();
-		}
-		else{
-			road=experiment.get(0).getExperimentVideopath();
-		}
+		String road = Constant.DIRECTORY.FILE;
+//		if (type.equals("PDF")) {
+//			road=experiment.get(0).getExperimentManualpath();
+//		}
+//		else{
+//			road=experiment.get(0).getExperimentVideopath();
+//		}
 		UploadUtils uploadUtils = new UploadUtils();
 		if (uploadUtils.uploadUtils(uploadfile, road)) {
 			fileservice.insert(f);
@@ -250,12 +256,14 @@ public class ExperimentController {
 			//改变文件的名字
 			if (type.equals("PDF")) {
 				if (fileUtil.updatename(road+"/"+uploadfile.getOriginalFilename(), road+"/"+uid+".pdf")) {
-					return ResultUtil.success("上传成功！");
+					ImageUtil.generatePDFImage(road+"/"+uid+".pdf", road+"/"+uid+".jpg");
+					return ResultUtil.success(uid);
 				}
 			}
 			else {
 				if (fileUtil.updatename(road+"/"+uploadfile.getOriginalFilename(), road+"/"+uid+".mp4")) {
-					return ResultUtil.success("上传成功！");
+					ImageUtil.generateVideoImage(Constant.FFMPEG_PATH, road+"/"+uid+".mp4", road+"/"+uid+".jpg");
+					return ResultUtil.success(uid);
 				}
 			}
 		}
@@ -278,14 +286,17 @@ public class ExperimentController {
 		}
 		String road="";
 		if (file.getFileType().equals("PDF")) {
-			 road = Constant.DIRECTORY.EXPERIMENT_REPORT+experimentId+"/"+fileId+".pdf";
+//			 road = Constant.DIRECTORY.EXPERIMENT_REPORT+experimentId+"/"+fileId+".pdf";
+			road = Constant.DIRECTORY.FILE+fileId+".pdf";
 		}
 		else{
-			 road = Constant.DIRECTORY.EXPERIMENT_VIDEO+experimentId+"/"+fileId+".mp4";
+//			 road = Constant.DIRECTORY.EXPERIMENT_VIDEO+experimentId+"/"+fileId+".mp4";
+			road = Constant.DIRECTORY.FILE+fileId+".mp4";
 		}
 		java.io.File test = new java.io.File(road);
-		if (test.isFile()) {
-			if (test.delete()) {
+		java.io.File pic = new java.io.File(Constant.DIRECTORY.FILE+fileId+".jpg");
+		if (test.isFile() && pic.isFile()) {
+			if (test.delete() && pic.delete()) {
 				if (fileservice.delete(file.getFileId())==1) {
 					return ResultUtil.success("删除成功！");
 				}
@@ -305,10 +316,16 @@ public class ExperimentController {
 		}
 		if (experimentFile.size()!=0) {
 			//删除目录
-			String road1 = Constant.DIRECTORY.EXPERIMENT_REPORT+experimentId;
-			String road2 = Constant.DIRECTORY.EXPERIMENT_VIDEO+experimentId;
-			fileUtil.deleteSection(road1);
-			fileUtil.deleteSection(road2);
+//			String road1 = Constant.DIRECTORY.EXPERIMENT_REPORT+experimentId;
+//			String road2 = Constant.DIRECTORY.EXPERIMENT_VIDEO+experimentId;
+			for(File f : experimentFile){
+				String type = "PDF".equals(f.getFileType()) ? ".pdf" : ".mp4";
+				String road1 = Constant.DIRECTORY.FILE + f.getFileId() + type;
+				String road2 = Constant.DIRECTORY.FILE + f.getFileId() + ".jpg";
+				fileUtil.deleteSection(road1);
+				fileUtil.deleteSection(road2);
+			}
+
 				
 		}
 		if (experimentService.deleteExperiment(experimentId, courseId)==0) {
